requirements.txt
black==22.3.0
cfgv==3.3.1
click==8.1.3
cycler==0.11.0
distlib==0.3.4
filelock==3.7.1
fonttools==4.33.3
identify==2.5.1
kiwisolver==1.4.3
matplotlib==3.5.2
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
virtualenv==20.14.1
